build: add ap mode compile
Summary:

Test Plan: run cramfs ramdisk ok

Reviewers: huiliang.xu, xiaofeng.ling, lei.mao

Reviewed By: huiliang.xu, xiaofeng.ling

Differential Revision: https://cr.hobot.cc/D14462

+++ build.sh
@@ -7,25 +7,23 @@ function choose()
     local rootfscpio="CONFIG_INITRAMFS_SOURCE=\"./usr/rootfs.cpio\""
     local rootfspre="CONFIG_INITRAMFS_SOURCE=\"./usr/prerootfs/\""
     local rootfsnone="CONFIG_INITRAMFS_SOURCE=\"\""
+    local manifest=$KERNEL_INITRAMFS_MANIFEST
     cp .config $conftmp
 
     if ! $hascpio ;then
         sed -i "s#${rootfscpio}#${rootfsnone}#g" $conftmp
     else
         sed -i "s#${rootfscpio}#${rootfspre}#g" $conftmp
-        rm -rf ${SRC_KERNEL_DIR}/usr/prerootfs/
-        mkdir -p ${SRC_KERNEL_DIR}/usr/prerootfs/
-        if [ "$BOOT_MODE" = "nor" ];then
-            export KERNEL_INITRAMFS_MANIFEST="$SRC_DEVICE_DIR/$TARGET_VENDOR/$TARGET_PROJECT/debug-kernel-rootfs.manifest"
+        if [ "$BOOT_MODE" != "ap" ];then
+            rm -rf ${SRC_KERNEL_DIR}/usr/prerootfs/
+            mkdir -p ${SRC_KERNEL_DIR}/usr/prerootfs/
+            if [ "$BOOT_MODE" = "nor" ];then
+                manifest="$SRC_DEVICE_DIR/$TARGET_VENDOR/$TARGET_PROJECT/debug-kernel-rootfs.manifest"
+            fi
+            ${SRC_SCRIPTS_DIR}/build_root_manifest.sh $manifest \
+                ${TARGET_PREROOTFS_DIR} ${SRC_KERNEL_DIR}/usr/prerootfs/
+            sed -i "/AMA0/d" ${SRC_KERNEL_DIR}/usr/prerootfs/etc/inittab
         fi
-        ${SRC_SCRIPTS_DIR}/build_root_manifest.sh ${KERNEL_INITRAMFS_MANIFEST} ${TARGET_PREROOTFS_DIR} ${SRC_KERNEL_DIR}/usr/prerootfs/
-        if [ ! -f "${SRC_KERNEL_DIR}/usr/prerootfs/init" ];then
-            echo "#!/bin/sh" > ${SRC_KERNEL_DIR}/usr/prerootfs/init
-            echo "exec /sbin/init \"\$@\"" >> ${SRC_KERNEL_DIR}/usr/prerootfs/init
-            chmod +x ${SRC_KERNEL_DIR}/usr/prerootfs/init
-        fi
-
-        sed -i "/AMA0/d" ${SRC_KERNEL_DIR}/usr/prerootfs/etc/inittab
     fi
 
     cp $conftmp .config
@@ -52,16 +50,6 @@ function make_recovery_img()
 
     local kernel_initram="$SRC_DEVICE_DIR/$TARGET_VENDOR/$TARGET_PROJECT/debug-kernel-rootfs.manifest"
     ${SRC_SCRIPTS_DIR}/build_root_manifest.sh ${kernel_initram} ${TARGET_PREROOTFS_DIR} ${SRC_KERNEL_DIR}/usr/prerootfs/
-    if [ ! -f "${SRC_KERNEL_DIR}/usr/prerootfs/init" ];then
-        echo "#!/bin/sh" > ${SRC_KERNEL_DIR}/usr/prerootfs/init
-        echo "/bin/mount -t devtmpfs devtmpfs /dev" >> ${SRC_KERNEL_DIR}/usr/prerootfs/init
-        echo "exec 0</dev/console" >> ${SRC_KERNEL_DIR}/usr/prerootfs/init
-        echo "exec 1>/dev/console" >> ${SRC_KERNEL_DIR}/usr/prerootfs/init
-        echo "exec 2>/dev/console" >> ${SRC_KERNEL_DIR}/usr/prerootfs/init
-        echo "exec /sbin/init \"\$@\"" >> ${SRC_KERNEL_DIR}/usr/prerootfs/init
-        chmod +x ${SRC_KERNEL_DIR}/usr/prerootfs/init
-    fi
-
     sed -i "/AMA0/d" ${SRC_KERNEL_DIR}/usr/prerootfs/etc/inittab
 
     cp $conftmp .config
